+++ src/helpers/types.js
@@ -82,8 +82,12 @@ export function attribute2object (value) {
         .join(',');
       return JSON.parse(`{${ normalized }}`);
     } catch (err) {
-      console.error(err);
-      return undefined;
+      try {
+        return str2value(value);
+      } catch (err) {
+        console.error(err);
+        return undefined;
+      }
     }
   } else if (isObject(value)) {
     return value;
@@ -224,7 +228,8 @@ function removeDoubleQuote (str) {
  * @returns {boolean}
  */
 export function isLikeObject (str) {
-  return /^\s*{(.|\s)*}\s*$/.test(str);
+  // // Catastrophic backtracking with /^\s*{(.|\s)*}\s*$/.test(str);
+  return /^\s*{/.test(str) && /}\s*$/.test(str);
 }
 
 /**
@@ -233,16 +238,24 @@ export function isLikeObject (str) {
  * @returns {boolean}
  */
 export function isLikeArray (str) {
-  return /^\s*\[(.|\s)*]\s*$/.test(str);
+  // Catastrophic backtracking with /^\s*\[(.|\s)*]\s*$/.test(str);
+  return /^\s*\[/.test(str) && /]\s*$/.test(str);
 }
 
 
 export function csvStr2obj (str) {
-  let keys = [];
-  return str
+  let keys    = [];
+  const parts = str
     .split(/(\r\n|\r|\n)/)
     .map(r => r.trim())
-    .filter(r => r)
+    .filter(r => r);
+  if (parts.length === 0) {
+    return [];
+  }
+  if (parts.length === 1) {
+    return str2value(parts[0]);
+  }
+  return parts
     .reduce(
       (result, row, idx) => {
         const obj   = {};