+++ test/component/composer/cases/case72.js
@@ -1,0 +1,10 @@
+export const title       = '72) datum: attribute (string)';
+export const description = `data="hello"`;
+
+export default `<g-composer data="hello">
+  <svg viewBox="0 0 110 110">
+    <text x="50" y="50" dominant-baseline="middle" text-anchor="middle"
+          g-content="data">test</text>
+  </svg>
+</g-composer>
+`;