+++ test/component/composer/cases/case76.js
@@ -1,0 +1,10 @@
+export const title       = '76) datum: data-src (string)';
+export const description = `data-src="datum-str.txt"`;
+
+export default `<g-composer data-src="/test/component/composer/assets/datum-str.txt">
+  <svg viewBox="0 0 110 110">
+    <text x="50" y="50" dominant-baseline="middle" text-anchor="middle"
+          g-content="data">test</text>
+  </svg>
+</g-composer>
+`;